pyproject.toml
[tool.poetry]
name = "casana"
version = "0.1.0"
description = "Peak Detection Neural Network Interview Question"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.0"
pyyaml = "^6.0.1"
plotly = "^5.19.0"
kaleido = "0.2.1"
tqdm = "^4.67.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"